requirements.txt
flask==3.0.3
gunicorn==21.2.0
scikit-learn==1.6.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
pillow
pytesseract